@@ -3,36 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "14.x"
+    "node": "20.x"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@mui/material": "^5.0.4",
-    "@supabase/supabase-js": "^1.24.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "bootstrap": "5.1.1",
-    "date-fns": "^2.25.0",
-    "final-form": "^4.20.2",
-    "final-form-material-ui": "^0.3.0",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.15.10",
+    "@mui/icons-material": "^5.15.10",
+    "@mui/lab": "^5.0.0-alpha.165",
+    "@supabase/supabase-js": "^2.39.3",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "bootstrap": "^5.3.3",
+    "date-fns": "^3.3.1",
+    "final-form": "^4.20.7",
     "material-table": "^1.69.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-rc.0",
-    "react-data-grid": "5.0.1",
-    "react-datepicker": "^4.2.1",
-    "react-dom": "^17.0.2",
-    "react-final-form": "^6.5.3",
-    "react-final-form-file-field": "^0.0.6",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.1",
-    "web-vitals": "^1.0.1"
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.10.1",
+    "react-data-grid": "^7.0.0-beta.2",
+    "react-datepicker": "^4.25.0",
+    "react-dom": "^18.2.0",
+    "react-final-form": "^6.5.9",
+    "react-router-dom": "^6.22.1",
+    "react-scripts": "5.0.1",
+    "styled-components": "^6.1.8",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
